Review

+++ task-impl/src/main/java/com/chutneytesting/task/jms/consumer/ConsumerFactory.java
@@ -24,7 +24,7 @@ public class ConsumerFactory {
         final Consumer consumer;
         if (arguments.bodySelector == null) {
             MessageConsumer messageConsumer = session.createConsumer(destination, arguments.selector);
-            consumer = new SimpleMessageConsumer(messageConsumer, parseToMs(arguments.timeout));
+            consumer = new SimpleMessageConsumer(messageConsumer, (int) parseToMs(arguments.timeout));
         } else {
             QueueBrowser browser = session.createBrowser((Queue) destination, arguments.selector);
             BodySelector bodySelector = bodySelectorFactory.build(arguments.bodySelector);
